github-agent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
coverage==7.8.0
cryptography==45.0.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
flake8==7.2.0
fsspec==2025.3.2
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
openai==1.79.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
portalocker==2.10.1
protobuf==6.31.0
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
PyGithub==2.6.1
PyJWT==2.10.1
pylint==3.3.7
PyNaCl==1.5.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dotenv==1.1.0
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
ruff==0.11.10
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
